==> 06_beta_diversity.Rmd <==
# Beta diversity

```{r load_data_beta, comment="", message=FALSE, warning=FALSE, eval=TRUE}
load("data/data.Rdata")
load("data/beta.Rdata")
```

#### Calculate beta Hill div

```{r beta_div, comment="", message=FALSE, warning=FALSE, eval=FALSE}
beta_q0n <- genome_counts_filt %>%
  column_to_rownames(., "genome") %>%
  filter(rowSums(. != 0, na.rm = TRUE) > 0) %>%
  select_if(~!all(. == 0)) %>%
  hillpair(., q = 0)

beta_q1n <- genome_counts_filt %>%
  column_to_rownames(., "genome") %>%
  filter(rowSums(. != 0, na.rm = TRUE) > 0) %>%
  select_if(~!all(. == 0)) %>%
  hillpair(., q = 1)

genome_counts_filt <- genome_counts_filt %>%
  column_to_rownames(., "genome") %>%
  filter(rowSums(. != 0, na.rm = TRUE) > 0) %>%
  select_if(~!all(. == 0))%>%
  rownames_to_column(., "genome")

genome_tree <- keep.tip(genome_tree, tip=genome_counts_filt$genome)

# beta_q1p <- genome_counts_filt %>%
#   column_to_rownames(., "genome") %>%
#   filter(rowSums(. != 0, na.rm = TRUE) > 0) %>%
#   select_if(~!all(. == 0)) %>%
#   hillpair(., q = 1, tree = genome_tree)

genome_counts_filt <- genome_counts[genome_counts$genome %in% rownames(genome_gifts),]
genome_counts_filt <- genome_counts_filt %>%
  remove_rownames() %>% 
  column_to_rownames(., "genome") %>%
  filter(rowSums(. != 0, na.rm = TRUE) > 0) %>%
  select_if(~!all(. == 0))%>%
  rownames_to_column(., "genome")

genome_gifts1 <- genome_gifts[rownames(genome_gifts) %in% genome_counts_filt$genome,]
genome_gifts1 <- genome_gifts1[, colSums(genome_gifts1 != 0) > 0]

dist <- genome_gifts1 %>%
  to.elements(., GIFT_db) %>%
  traits2dist(., method = "gower")

# beta_q1f <- genome_counts_filt %>%
# #  remove_rownames() %>% 
#   column_to_rownames(., "genome") %>%
#   filter(rowSums(. != 0, na.rm = TRUE) > 0) %>%
#   select_if(~!all(. == 0)) %>%
#   hillpair(., q = 1, dist = dist)

```
#### Save outputs
```{r save_beta, comment="", message=FALSE,echo=FALSE,warning=FALSE, eval=FALSE}
save(beta_q0n, 
     beta_q1n, 
     # beta_q1p, 
     # beta_q1f,
     genome_counts,
     sample_metadata,
     genome_metadata,
     genome_gifts,
     genome_gifts1,
     genome_tree,
     file = "data/beta.Rdata")
```

### Permanova analysis
```{r}
set.seed(1234)
```

#### Richness
```{r permanova_q0, comment="", message=FALSE, warning=FALSE}
sample_metadata_row <- column_to_rownames(sample_metadata, "sample") 
sample_metadata_row <- sample_metadata_row[labels(beta_q0n$S), ]
sample_metadata_trim <- sample_metadata %>%
  filter(sample %in% rownames(sample_metadata_row))

betadisper(beta_q0n$S, sample_metadata_row$host_order) %>% permutest(., pairwise = TRUE) 

adonis2(beta_q0n$S ~ host_order, 
        data = sample_metadata_trim %>% arrange(match(sample,labels(beta_q0n$S))), 
        permutations = 999) %>%
        broom::tidy() %>%
        tt()
```

#### Neutral
```{r permanova_wildcap_q1, comment="", message=FALSE, warning=FALSE}
betadisper(beta_q1n$S, sample_metadata_row$host_order) %>% permutest(., pairwise = TRUE) 

adonis2(beta_q1n$S ~ host_order, 
        data = sample_metadata_trim %>% arrange(match(sample,labels(beta_q1n$S))), 
        permutations = 999) %>%
        broom::tidy() %>%
        tt()
```

#### Phylogenetic
```{r permanova_wildcap_qP, comment="", message=FALSE, warning=FALSE}
# betadisper(beta_q1p$S, sample_metadata_row$host_order) %>% permutest(., pairwise = TRUE) 
# adonis2(beta_q1p$S ~ host_order, 
#         data = sample_metadata %>% arrange(match(sample,labels(beta_q1p$S))), 
#         permutations = 999) %>%
#         broom::tidy() %>%
#         tt()
```

#### Functional
```{r permanova_wildcap_qF, comment="", message=FALSE, warning=FALSE}
# betadisper(beta_q1f$S, sample_metadata_row$host_order) %>% permutest(., pairwise = TRUE) 
# adonis2(beta_q1f$S ~ host_order,
#         data = sample_metadata %>% arrange(match(sample,labels(beta_q1f$S))),
#         permutations = 999) %>%
#         broom::tidy() %>%
#         tt()
```

### Beta diversity plots

#### Richness

#### Richness (host species)

```{r beta_div_nmds_richness_plot_median_species, comment="", message=FALSE, warning=FALSE, fig.height=7, fig.width=10, fig.fullwidth=TRUE, eval=TRUE}

ggplotly(

beta_q0n$S %>%
  vegan::metaMDS(., trymax = 500, k = 2, trace=0) %>%
  vegan::scores() %>%
  as_tibble(., rownames = "sample") %>%
  dplyr::left_join(sample_metadata, by = join_by(sample == sample)) %>%
  group_by(host_species) %>%
  mutate(x_cen = mean(NMDS1, na.rm = TRUE)) %>%
  mutate(y_cen = mean(NMDS2, na.rm = TRUE)) %>%
  ungroup() %>%
  ggplot(aes(x = NMDS1, y = NMDS2, color = host_species, shape = host_order,
             label = sample, label1 = singlem_fraction)) +
    geom_point(size = 4) +
    #   stat_ellipse(aes(color = beta_q1n_nmds$Groups))+
    geom_segment(aes(x = x_cen, y = y_cen, xend = NMDS1, yend = NMDS2), alpha = 0.9, show.legend = FALSE) +
    theme_classic() +
    theme(
      axis.text.x = element_text(size = 12),
      axis.text.y = element_text(size = 12),
      axis.title = element_text(size = 12, face = "bold"),
      axis.text = element_text(face = "bold", size = 12),
      panel.background = element_blank(),
      axis.line = element_line(size = 0.5, linetype = "solid", colour = "black"),
      legend.text = element_text(size = 12),
      legend.title = element_text(size = 14),
      legend.position = "none"
    )

) 
```

#### Neutral


#### Neutral (host species)

```{r beta_div_nmds_neutral_plot_median_species, comment="", message=FALSE, warning=FALSE, fig.height=7, fig.width=10, fig.fullwidth=TRUE, eval=TRUE}

ggplotly(

beta_q1n$S %>%
  vegan::metaMDS(., trymax = 500, k = 2, trace=0) %>%
  vegan::scores() %>%
  as_tibble(., rownames = "sample") %>%
  dplyr::left_join(sample_metadata, by = join_by(sample == sample)) %>%
  group_by(host_species) %>%
  mutate(x_cen = mean(NMDS1, na.rm = TRUE)) %>%
  mutate(y_cen = mean(NMDS2, na.rm = TRUE)) %>%
  ungroup() %>%
  ggplot(aes(x = NMDS1, y = NMDS2, color = host_species, shape = host_order,
             label = sample, label1 = singlem_fraction)) +
    geom_point(size = 4) +
    #   stat_ellipse(aes(color = beta_q1n_nmds$Groups))+
    geom_segment(aes(x = x_cen, y = y_cen, xend = NMDS1, yend = NMDS2), alpha = 0.9, show.legend = FALSE) +
    theme_classic() +
    theme(
      axis.text.x = element_text(size = 12),
      axis.text.y = element_text(size = 12),
      axis.title = element_text(size = 12, face = "bold"),
      axis.text = element_text(face = "bold", size = 12),
      panel.background = element_blank(),
      axis.line = element_line(size = 0.5, linetype = "solid", colour = "black"),
      legend.text = element_text(size = 12),
      legend.title = element_text(size = 14),
      legend.position = "none"
    )

)

```
